requirements.txt
customtkinter>=5.2.0
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
segyio>=1.9.10
